updates to include VS2017 build for win81
