+++ ADAL.NET.VS2015.sln
@@ -13,6 +13,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdalUniversalTestApp.Window
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ADAL.WinRT.VS2015.Platform", "src\Microsoft.IdentityModel.Clients.ActiveDirectory.Platform\ADAL.WinRT.VS2015.Platform.csproj", "{62659826-9AC1-4592-A1DB-1E7EFBDBD426}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Identity.Core.VS2015", "src\Microsoft.Identity.Core\Microsoft.Identity.Core.VS2015.csproj", "{72F03823-45DC-4F77-BDCB-29E880BF1AC8}"
+EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
 		devApps\AdalUniversalTestApp\AdalUniversalSample.Shared\AdalUniversalSample.Shared.projitems*{3e89ccea-0c4b-461e-95a6-a906e5583af9}*SharedItemsImports = 4
@@ -85,6 +87,22 @@ Global
 		{62659826-9AC1-4592-A1DB-1E7EFBDBD426}.Release|x64.Build.0 = Release|x64
 		{62659826-9AC1-4592-A1DB-1E7EFBDBD426}.Release|x86.ActiveCfg = Release|x86
 		{62659826-9AC1-4592-A1DB-1E7EFBDBD426}.Release|x86.Build.0 = Release|x86
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Debug|ARM.ActiveCfg = Debug|ARM
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Debug|ARM.Build.0 = Debug|ARM
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Debug|x64.ActiveCfg = Debug|x64
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Debug|x64.Build.0 = Debug|x64
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Debug|x86.ActiveCfg = Debug|x86
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Debug|x86.Build.0 = Debug|x86
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Release|ARM.ActiveCfg = Release|ARM
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Release|ARM.Build.0 = Release|ARM
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Release|x64.ActiveCfg = Release|x64
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Release|x64.Build.0 = Release|x64
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Release|x86.ActiveCfg = Release|x86
+		{72F03823-45DC-4F77-BDCB-29E880BF1AC8}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
