+++ src/Microsoft.Identity.Core/Microsoft.Identity.Core.VS2015.csproj
@@ -8,7 +8,7 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProductVersion>8.0.30703</ProductVersion>
     <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>{826D90F4-CA79-4788-918C-B66AA5CC6DF0}</ProjectGuid>
+    <ProjectGuid>{72F03823-45DC-4F77-BDCB-29E880BF1AC8}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>Microsoft.IdentityModel.Clients.ActiveDirectory</RootNamespace>
