+++ src/Microsoft.Identity.Core/Platforms/android/CryptographyHelper.cs
