+++ src/Microsoft.Identity.Core/Platforms/iOS/CryptographyHelper.cs
