+++ tests/Test.ADAL.NET.Unit/CommonUnitTests.cs
@@ -26,7 +26,7 @@
 //------------------------------------------------------------------------------
 
 using System;
-
+using Microsoft.Identity.Core;
 using Microsoft.IdentityModel.Clients.ActiveDirectory;
 using Microsoft.IdentityModel.Clients.ActiveDirectory.Internal.Helpers;
 using Microsoft.IdentityModel.Clients.ActiveDirectory.Internal.Platform;
